requirements.txt
Django==3.0.4
django-heroku==0.3.1
gunicorn==20.0.4
honeycomb-beeline==2.12.1
